Wordcloud 23/07/20
Axis no longer appear when wordcloud is saved, and it can now be opened. Added Prompt to ask the user for login details so they can enter their own and also gave them the opportunity to enter their own URL. The code has been cleared up a bit, but not completely. In process of getting Publication information so that the Wordcloud will show more than just the IDs.